+++ src/libtcod/fov_restrictive.c
@@ -234,10 +234,11 @@ static void compute_quadrant(
   }
 }
 
-void TCOD_map_compute_fov_restrictive_shadowcasting(
+TCOD_Error TCOD_map_compute_fov_restrictive_shadowcasting(
     TCOD_Map* __restrict map, int pov_x, int pov_y, int max_radius, bool light_walls) {
   if (!TCOD_map_in_bounds(map, pov_x, pov_y)) {
-    return;  // Invalid POV.
+    TCOD_set_errorvf("Point of view {%i, %i} is out of bounds.", pov_x, pov_y);
+    return TCOD_E_INVALID_ARGUMENT;
   }
   /* set PC's position as visible */
   map->cells[pov_x + (pov_y * map->width)].fov = true;
@@ -246,7 +247,12 @@ void TCOD_map_compute_fov_restrictive_shadowcasting(
   const int max_obstacles = map->nbcells / 7;
   double* start_angle = malloc(max_obstacles * sizeof(*start_angle));
   double* end_angle = malloc(max_obstacles * sizeof(*end_angle));
-
+  if (!start_angle || !end_angle) {
+    free(end_angle);
+    free(start_angle);
+    TCOD_set_errorv("Out of memory.");
+    return TCOD_E_OUT_OF_MEMORY;
+  }
   /* compute the 4 quadrants of the map */
   compute_quadrant(map, pov_x, pov_y, max_radius, light_walls, 1, 1, start_angle, end_angle);
   compute_quadrant(map, pov_x, pov_y, max_radius, light_walls, 1, -1, start_angle, end_angle);
@@ -255,4 +261,5 @@ void TCOD_map_compute_fov_restrictive_shadowcasting(
 
   free(end_angle);
   free(start_angle);
+  return TCOD_E_OK;
 }